Added Va input. Fixed vector orientation issues.
Currently requires vertical vector inputs for b and Va, might be able to make this more robust in the future - but this would allow us to be sloppy with vector orientation in other portions of the code. With an input of all zeros for b and 1V for every joint voltage input, I get out huge velocities. Probably needs to be fixed.